+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Add Billing API
+
 ## [2.7.0] - 2026-02-24 
 - Add Sandbox Messages API
 - Add Sending Domains API