+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Add Account Accesses API
+- Add Billing API
 
 ## [2.7.0] - 2026-02-24 
 - Add Sandbox Messages API